pyproject.toml
[tool.poetry]
name = "trado"
version = "0.1.1"
description = "Dumb and dirty docker-compose files generator for traefik exposing"
authors = ["Grigory Bakunov <thebobuk@ya.ru>"]
license = "The Unlicense"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.scripts]
trado = "trado:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
